# final_homework/api/bff/mobile/v1/mobile.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "mobile api",
    "version": "0.0.1"
  },
  "tags": [
    {
      "name": "Mobile"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/current_user": {
      "get": {
        "summary": "User",
        "operationId": "Mobile_CurrentUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CurrentUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Mobile"
        ],
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "operationId": "Mobile_UpdatePost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdatePostReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpdatePostRequest"
            }
          }
        ],
        "tags": [
          "Mobile"
        ],
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/posts": {
      "get": {
        "operationId": "Mobile_PostIndex",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListPostReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Mobile"
        ],
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "summary": "Post",
        "operationId": "Mobile_CreatePost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreatePostReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreatePostRequest"
            }
          }
        ],
        "tags": [
          "Mobile"
        ],
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/posts/{id}": {
      "get": {
        "operationId": "Mobile_PostShow",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetPostReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Mobile"
        ],
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/sign_in_by_pwd": {
      "post": {
        "operationId": "Mobile_SignInByPwd",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SignInResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SignInByPwdRequest"
            }
          }
        ],
        "tags": [
          "Mobile"
        ]
      }
    },
    "/api/sign_up": {
      "post": {
        "summary": "Auth",
        "operationId": "Mobile_SignUp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SignUpResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SignUpRequest"
            }
          }
        ],
        "tags": [
          "Mobile"
        ]
      }
    }
  },
  "definitions": {
    "ListPostReplyPost": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "published": {
          "type": "boolean"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "typeUrl": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreatePostReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "published": {
          "type": "boolean"
        }
      }
    },
    "v1CreatePostRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "published": {
          "type": "boolean"
        }
      },
      "title": "Post"
    },
    "v1CurrentUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "phone": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        }
      }
    },
    "v1GetPostReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "published": {
          "type": "boolean"
        }
      }
    },
    "v1ListPostReply": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ListPostReplyPost"
          }
        }
      }
    },
    "v1SignInByPwdRequest": {
      "type": "object",
      "properties": {
        "phone": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "v1SignInResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
    "v1SignUpRequest": {
      "type": "object",
      "properties": {
        "phone": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "v1SignUpResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "phone": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        }
      }
    },
    "v1UpdatePostReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1UpdatePostRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "published": {
          "type": "boolean"
        }
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}